criação das validações e handler

+++ desafio3-CRUD/desafiocrud/src/main/java/com/devsuperior/felipe/desafiocrud/services/ClientService.java
@@ -4,11 +4,14 @@ import com.devsuperior.felipe.desafiocrud.entities.Client;
 import com.devsuperior.felipe.desafiocrud.entities.dto.ClientDTO;
 import com.devsuperior.felipe.desafiocrud.repositories.ClientRepository;
 
+import com.devsuperior.felipe.desafiocrud.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeMap;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -32,7 +35,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = repository.findById(id).orElseThrow();
+        Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não Encontrado"));
         return mapper.map(client, ClientDTO.class);
     }
 
@@ -50,14 +53,20 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client client = repository.getReferenceById(id);
-        this.mapper.map(dto,client);
-        Client updatedClient = repository.save(client);
-        return mapper.map(updatedClient, ClientDTO.class);
+        /* usando o mapper é gerado um erro de nível mais baixo, semelhante ao erro gerado
+         com Constraint Violation no exemplo dado pelo professor, por isso realizei a checagem existsById
+         para lançar manualmente a exceção
+         */
+        if (!repository.existsById(id)) throw new ResourceNotFoundException("Recurso não Encontrado");
+            Client client = repository.getReferenceById(id);
+            this.mapper.map(dto, client);
+            Client updatedClient = repository.save(client);
+            return mapper.map(updatedClient, ClientDTO.class);
     }
 
     @Transactional
     public void deleteById(Long id) {
+        if (!repository.existsById(id)) throw new ResourceNotFoundException("Recurso não Encontrado");
         repository.deleteById(id);
     }
 }
